# docs/contracts/J21-J23.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DAS J21-J23 additive API draft",
    "version": "0.1.0-proposed",
    "description": "Design-only as of 2026-10-08. All five operations below are proposals; no production endpoint or authority integration is enabled by this file."
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Local preview only; these draft endpoints have not been installed."
    }
  ],
  "paths": {
...
    "/api/v2/admin/catalogs": {
      "get": {
        "operationId": "listAdminCatalogs",
        "summary": "J22 management list including inactive catalogs",
        "tags": [
          "Document"
        ],
        "parameters": [
          {
            "name": "group",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "companies",
                "methods",
                "documentTypes",
                "internalTypes",
                "sensitivity",
                "categories"
              ]
            }
          },
          {
            "name": "activity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "Active",
                "Inactive",
                "All"
              ],
              "default": "Active"
            }
          },
          {
            "name": "searchTerm",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000000,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "PROPOSED, not implemented. Unknown/duplicate/blank query values rejected. Read docs/contracts for scope validation, paging invariants, error codes and rollout.",
        "responses": {
          "200": {
            "description": "Valid response. For My Staff, may be explicitly degraded with tasks=null.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogAdminResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "group": "methods",
                    "activity": "Inactive",
                    "canEditGroup": true,
                    "items": [
                      {
                        "id": "33333333-3333-4333-8333-333333333333",
                        "group": "methods",
                        "code": "QA_METHOD",
                        "name": "Phương thức kiểm thử",
                        "sortOrder": 1,
                        "isActive": false,
                        "version": 2
                      }
                    ],
                    "totalCount": 1,
                    "pageNumber": 1,
                    "pageSize": 20
                  },
                  "message": null,
                  "errors": [],
                  "traceId": "qa-example"
                }
              }
            }
          },
          "400": {
            "description": "Error; no stale or partial business data returned.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated; middleware may return an empty challenge body.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogError"
                }
              }
            }
          },
          "403": {
            "description": "Error; no stale or partial business data returned.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogError"
                }
              }
            }
          },
          "503": {
            "description": "Error; no stale or partial business data returned.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/admin/catalogs/options": {
      "get": {
        "operationId": "getCatalogAdminOptions",
        "summary": "J22 effective management capability metadata",
        "tags": [
          "Document"
        ],
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "PROPOSED, not implemented. Unknown/duplicate/blank query values rejected. Read docs/contracts for scope validation, paging invariants, error codes and rollout.",
        "responses": {
          "200": {
            "description": "Valid response. For My Staff, may be explicitly degraded with tasks=null.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogOptionsResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "canManage": false
                  },
                  "message": null,
                  "errors": [],
                  "traceId": "qa-example"
                }
              }
            }
          },
          "400": {
            "description": "Error; no stale or partial business data returned.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated; middleware may return an empty challenge body.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogError"
                }
              }
            }
          },
          "403": {
            "description": "Error; no stale or partial business data returned.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogError"
                }
              }
            }
          },
          "503": {
            "description": "Error; no stale or partial business data returned.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CatalogItem": {
        "type": "object",
        "required": [
          "id",
          "group",
          "code",
          "name",
          "sortOrder",
          "isActive",
          "version"
        ],
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^(?!00000000-0000-0000-0000-000000000000$)[0-9a-fA-F]{8}(?:-[0-9a-fA-F]{4}){3}-[0-9a-fA-F]{12}$"
          },
          "group": {
            "type": "string",
            "enum": [
              "companies",
              "methods",
              "documentTypes",
              "internalTypes",
              "sensitivity",
              "categories"
            ]
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_]{1,64}$"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "sortOrder": {
            "type": "integer",
            "minimum": 0,
            "maximum": 10000
          },
          "isActive": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          }
        }
      },
      "CatalogAdminPage": {
        "type": "object",
        "required": [
          "items",
          "totalCount",
          "pageNumber",
          "pageSize",
          "group",
          "activity",
          "canEditGroup"
        ],
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/CatalogItem"
            }
          },
          "totalCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "pageNumber": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000000
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "group": {
            "type": "string",
            "enum": [
              "companies",
              "methods",
              "documentTypes",
              "internalTypes",
              "sensitivity",
              "categories"
            ]
          },
          "activity": {
            "type": "string",
            "enum": [
              "Active",
              "Inactive",
              "All"
            ]
          },
          "canEditGroup": {
            "type": "boolean"
          }
        }
      },
      "CatalogAdminResponse": {
        "type": "object",
        "required": [
          "success",
          "data",
          "message",
          "errors",
          "traceId"
        ],
        "additionalProperties": false,
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/CatalogAdminPage"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "maxItems": 0,
            "items": {
              "$ref": "#/components/schemas/PartnerErrorEntry"
            }
          },
          "traceId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        }
      },
      "CatalogOptionsResponse": {
        "type": "object",
        "required": [
          "success",
          "data",
          "message",
          "errors",
          "traceId"
        ],
        "additionalProperties": false,
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "type": "object",
            "required": [
              "canManage"
            ],
            "additionalProperties": false,
            "properties": {
              "canManage": {
                "type": "boolean"
              }
            }
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "maxItems": 0,
            "items": {
              "$ref": "#/components/schemas/PartnerErrorEntry"
            }
          },
          "traceId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        }
      },
...
      "PartnerErrorEntry": {
        "type": "object",
        "required": [
          "field",
          "code",
          "message"
        ],
        "additionalProperties": false,
        "properties": {
          "field": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "CatalogError": {
        "type": "object",
        "required": [
          "success",
          "data",
          "message",
          "errors",
          "traceId"
        ],
        "additionalProperties": false,
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "data": {
            "type": "object",
            "nullable": true,
            "enum": [
              null
            ]
          },
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/PartnerErrorEntry"
            }
          },
          "traceId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        }
      }
    }
  }
}